[improvement][agent] Extract the max_loops="auto" loop into AutonomousAgentLoop
agent.py was 5,237 lines, and roughly a fifth of it was the autonomous
plan-execute-summarize loop -- code that never runs unless
max_loops == "auto". Move it to swarms/agents/autonomous_loop.py.

agent.py: 5,224 -> 4,084 lines (-1,140).

Moved (1,164 lines), chosen by a fixpoint over the call graph: every method
whose referrers are all inside the cluster and which has zero references
outside agent.py:

  _run_autonomous_loop           862
  _create_plan_tool               94
  _subtask_done_tool              90
  _think_tool                     79
  _get_next_executable_subtask    25
  _all_subtasks_complete          14

AutonomousAgentLoop mirrors the existing LLMManager arrangement -- it takes
the owning agent and reads/writes agent state through that reference, so
`self.X` inside the moved code became `self.agent.X`. Agent constructs it
beside the other managers and keeps `_run_autonomous_loop` as a documented
delegate, so `max_loops="auto"` and any external caller are unaffected.
18 imports that moved with the code were pruned from agent.py.

Deliberately left on Agent, because they are not exclusive to the loop:
  - _generate_final_summary (13 external references; a streaming test calls it)
  - _complete_task_tool (also reached from execute_tools, which _run uses)

Tests: new TestAutonomousAgentLoop in tests/structs/test_agent.py (7 tests)
covering the wiring, that Agent delegates rather than reimplements, that
max_loops="auto" routes to the loop while a fixed count does not, dependency
gating and completion logic in the two pure helpers, and an end-to-end run
against a stubbed LLM asserting the plan lands on the agent.

Verified behaviour-identical to before the move: the same stubbed autonomous
run produces the same LLM call count, prompt sequence, and output length.
